kubevirt: upgrade k3s,multus,kubevirt,cdi,longhorn
A kube container generation number is tracked in cluster-update.sh
and compared to an 'applied' version in the persistent node fs /var/lib.
When the applied version is behind it triggers component upgrades.
cluster-update.sh uses an inline command in the zedkube micro service to publish
status updates to zedagent which will trigger info messages (ZInfoKubeClusterUpdateStatus) to a controller.

Fix: move baseosmgr initiated drain to config.Activated && !status.Activated.
	previously it could request drain for baseimage re-downloads without activation.

Signed-off-by: Andrew Durbin <andrewd@zededa.com>